Refactor FastCommitManager to DefaultCommitManager (#15)
* Refactor FastCommitManager to DefaultCommitManager

Renamed FastCommitManager to DefaultCommitManager throughout the codebase:
- Renamed src/FastCommitManager.sol to src/DefaultCommitManager.sol
- Renamed test/FastCommitManagerTest.sol to test/DefaultCommitManagerTest.sol
- Updated all imports and references in test files
- Updated script/EngineAndPeriphery.sol
- Updated all error selector references



* format

---------

+++ src/DefaultRuleset.sol
@@ -11,14 +11,13 @@ import {IRuleset} from "./IRuleset.sol";
 import {IEffect} from "./effects/IEffect.sol";
 
 contract DefaultRuleset is IRuleset {
-
     IEngine immutable ENGINE;
 
     IEffect[] public effects;
 
     constructor(IEngine _ENGINE, IEffect[] memory _effects) {
         ENGINE = _ENGINE;
-        for (uint i; i < _effects.length; i++) {
+        for (uint256 i; i < _effects.length; i++) {
             effects.push(_effects[i]);
         }
     }
